# ctsr-package/api/openapi.yaml
openapi: 3.1.0
info:
  title: Clinical Trial Systems Register (CTSR) API
  description: |
    REST API for the Clinical Trial Systems Register - maintaining inventory of 
    computerized systems used in clinical trials per ICH E6(R3) requirements.
    
    ## Authentication
    All endpoints require Bearer token authentication via Azure AD / Entra ID.
  version: 1.1.0
  contact:
    name: R&D IT Architecture
servers:
  - url: https://ctsr-dev.novonordisk.com/api/v1
    description: Development
  - url: https://ctsr.novonordisk.com/api/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [Vendors]
      summary: List vendors
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: vendor_type
          in: query
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Vendor list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorListResponse'

    post:
      operationId: createVendor
      tags: [Vendors]
      summary: Create vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorCreate'
      responses:
        '201':
          description: Vendor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
  /vendors/{vendor_id}:
    get:
      operationId: getVendor
      tags: [Vendors]
      summary: Get vendor details
      parameters:
        - $ref: '#/components/parameters/vendorIdParam'
      responses:
        '200':
          description: Vendor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateVendor
      tags: [Vendors]
      summary: Update vendor
      parameters:
        - $ref: '#/components/parameters/vendorIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorUpdate'
      responses:
        '200':
          description: Vendor updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vendor'

  # ===========================================================================
  # Systems (4 endpoints)
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    vendorIdParam:
      name: vendor_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # -------------------------------------------------------------------------
    # Lookups
    # -------------------------------------------------------------------------
...
    Vendor:
      type: object
      properties:
        vendor_id:
          type: string
          format: uuid
        vendor_code:
          type: string
        vendor_name:
          type: string
        vendor_type:
          type: string
        contact_name:
          type: string
        contact_email:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    VendorCreate:
      type: object
      required: [vendor_code, vendor_name, vendor_type]
      properties:
        vendor_code:
          type: string
        vendor_name:
          type: string
        vendor_type:
          type: string
        contact_name:
          type: string
        contact_email:
          type: string
    VendorUpdate:
      type: object
      properties:
        vendor_name:
          type: string
        vendor_type:
          type: string
        contact_name:
          type: string
        contact_email:
          type: string
        is_active:
          type: boolean
    VendorListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Vendor'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # -------------------------------------------------------------------------
    # Systems
    # -------------------------------------------------------------------------
...